vdesk: Unify type as unsigned
In some places vdesk was unsigned long, others unsigned int.  Convert
all to unsigned int and add conversion for ewmh vdesk property.

+++ evilwm.h
@@ -155,7 +155,7 @@ struct Client {
 	int             win_gravity;
 	int             old_border;
 #ifdef VWM
-	unsigned long   vdesk;
+	unsigned int vdesk;
 #endif
 	int             is_dock;
 	int             remove;  /* set when client needs to be removed */
@@ -170,7 +170,7 @@ struct Application {
 	unsigned int width, height;
 	int is_dock;
 #ifdef VWM
-	unsigned long vdesk;
+	unsigned int vdesk;
 #endif
 };
